fix(dbManager): fix racing condition between closing and auto-opening the DB
When going back from the root group, DB closing request is handled asynchronously.
So, UnlockDatabaseVC can appear before the DB is actually closed.
If the database key is remembered in keychain, UnlockDatabaseVC could start
auto-reopening the DB simultaneously with the ongoing closing.

So we clear the stored key synchronously when closing, to avoid such racing.